+++ app/routes/users/__tests__/register.test.ts
@@ -1,9 +1,9 @@
 import { tinyassert } from "@hiogawa/utils";
 import { beforeEach, describe, expect, it } from "vitest";
 import { T, db } from "../../../db/drizzle-client.server";
+import { testLoader } from "../../../misc/test-helper";
 import { findByUsername, getSessionUser } from "../../../utils/auth";
 import { getSession } from "../../../utils/session.server";
-import { testLoader } from "../../__tests__/helper";
 import { action } from "../register";
 
 describe("register.action", () => {
